Export styling into styled components
Use the GameCard component
