--- src/components/game-card/game-card.css
@@ -1,25 +1,0 @@
-.game-card {
-  position: relative;
-  width: 300px;
-  height: 300px;
-  background-size: cover;
-  background-position: center;
-  overflow: hidden;
-}
-
-.game-card::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: radial-gradient(circle at center, transparent, black);
-}
-
-.content {
-  position: relative;
-  background: white;
-  padding: 20px;
-  box-sizing: border-box;
-}
